Cleared superfluous files. Created Ork1 and Knight1 spritesheets. Created Animations for ORK1 and KNIGHT1.
Finished creating the sprite sheets for ORK1 and KNIGHT1. Used half-scale variants to reduce necessity to scale in the game screen. Removed all the superfluous files (eps, ai...etc.) from the sprite packages. Set up the animations for ORK1 and KNIGHT1 and set the default animation for KNIGHT1 to IDLE and the default animation of ORK1 to ORK1_RUN so that I can set it up to run onto the screen later.