Asset reorganization, CharacterManager script, Character script update, Accuracy indicators, and more.
Reorganized animators and animations. Created indicators for accuracy of taps. Created prefab to manage those indicators. Decided to use Sprite indicators instead of text because that was easier to position. Created prefabs for Knight1 and Ork1. Created CharacterManager class, which is just used for easy reference to various character prefabs based on their string name in a dictionary for fast lookup. Moved 'combat' logic to the CombatController, including TapTarget spawning. Brought in a simple script that should manage camera sizing on different sized devices.